move reactless functions of page content component to service
