+++ components/ClassPage/PageContent.tsx
@@ -2,15 +2,15 @@ import { mean } from 'lodash';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { ReactElement } from 'react';
-import { GetClassPageInfoQuery } from '../../generated/graphql';
 import {
   creditsDescription,
   creditsNumericDisplay,
 } from '../common/CreditsDisplay';
 import { LastUpdated } from '../common/LastUpdated';
-import { getCampusByLastDigit, getSeason, getYear } from '../global';
 import { isCompositeReq } from '../ResultsPage/Results/useResultDetail';
-import { Campus, CompositeReq, CourseReq, Requisite } from '../types';
+import { CompositeReq, CourseReq, Requisite } from '../types';
+import { GetClassPageInfoQuery } from '../../generated/graphql';
+import PageContentService from './PageContentService';
 
 type PageContentProps = {
   termId: string;
@@ -21,8 +21,21 @@ type PageContentProps = {
   isCoreq: boolean;
 };
 
+type ClassPageReqsBodyProps = {
+  termId: string;
+  campus: string;
+  classPageInfo: GetClassPageInfoQuery;
+};
+
 type ClassPageInfoProp = {
   classPageInfo: GetClassPageInfoQuery;
+};
+
+type PrereqsDisplayProps = {
+  termId: string;
+  campus: string;
+  prereqs: Requisite;
+  indents: number;
 };
 
 export default function PageContent({
@@ -116,36 +129,47 @@ function ClassPageInfoBody({ classPageInfo }: ClassPageInfoProp): ReactElement {
         <HeaderBody header="COURSE DESCRIPTION" body={latestOccurrence.desc} />
         <HeaderBody
           header="COURSE LEVEL"
-          body={getCourseLevel(latestOccurrence.termId.toString())}
+          body={PageContentService.getCourseLevel(
+            latestOccurrence.termId.toString()
+          )}
         />
       </div>
       <div className="verticalLine" />
       <div className="classPageBodyRight">
         <HeaderBody
           header="RECENT PROFESSORS"
-          body={getProfessors(classPageInfo, 10).join(', ')}
+          body={PageContentService.getProfessors(classPageInfo, 10).join(', ')}
         />
         <HeaderBody
           header="RECENT SEMESTERS"
-          body={getRecentSemesterNames(classPageInfo, 6).join(', ')}
+          body={PageContentService.getRecentSemesterNames(
+            classPageInfo,
+            6
+          ).join(', ')}
         />
         <div className="flex justify-space-between">
           <HeaderBody
             className="lg-text avgSeatsFilled"
             header="AVG SEATS FILLED"
-            body={`${Math.round(mean(seatsFilled(classPageInfo)))}`}
+            body={`${Math.round(
+              mean(PageContentService.seatsFilled(classPageInfo))
+            )}`}
           />
           <HeaderBody
             className="lg-text avgSeatsAvail"
             header="AVG SEATS AVAILABLE"
-            body={`${Math.round(mean(seatsAvailable(classPageInfo)))}`}
+            body={`${Math.round(
+              mean(PageContentService.seatsAvailable(classPageInfo))
+            )}`}
           />
         </div>
         <div className="flex justify-space-between">
           <HeaderBody
             className="lg-text avgNumSections"
             header="AVG # SECTIONS"
-            body={`${Math.round(mean(numberOfSections(classPageInfo)))}`}
+            body={`${Math.round(
+              mean(PageContentService.numberOfSections(classPageInfo))
+            )}`}
           />
           <HeaderBody
             className={`lg-text courseFees ${
@@ -163,12 +187,6 @@ function ClassPageInfoBody({ classPageInfo }: ClassPageInfoProp): ReactElement {
     </div>
   );
 }
-
-type ClassPageReqsBodyProps = {
-  termId: string;
-  campus: string;
-  classPageInfo: GetClassPageInfoQuery;
-};
 
 function ClassPageReqsBody({
   termId,
@@ -258,82 +276,6 @@ function HeaderBody({
     </div>
   );
 }
-
-function getCourseLevel(termId: string): string {
-  const termIdLastDigit = termId.charAt(termId.length - 1);
-  const campus = getCampusByLastDigit(termIdLastDigit);
-  return campus === Campus.NEU ? 'Undergraduate' : 'Graduate';
-}
-
-function getProfessors(
-  classPageInfo: GetClassPageInfoQuery,
-  limit: number
-): string[] {
-  const allOccurrences = classPageInfo.class.allOccurrences;
-  const profs = new Set<string>();
-  // at least display all the professors for the most recent occurrence of this course
-  allOccurrences[0].sections.forEach((section) => {
-    section.profs.forEach((prof) => profs.add(prof));
-  });
-  if (profs.size >= limit || allOccurrences.length <= 1)
-    return Array.from(profs);
-  else {
-    allOccurrences.slice(1).forEach((occurrence) => {
-      occurrence.sections.forEach((section) =>
-        section.profs.forEach((prof) => profs.add(prof))
-      );
-    });
-    return Array.from(profs).slice(0, limit);
-  }
-}
-
-function getRecentSemesterNames(
-  classPageInfo: GetClassPageInfoQuery,
-  limit: number
-): string[] {
-  const allSemesters = classPageInfo.class.allOccurrences.map((occurrence) => {
-    const termId = occurrence.termId.toString();
-    return `${getSeason(termId)} ${getYear(termId)}`;
-  });
-  return allSemesters.slice(0, Math.min(limit, allSemesters.length));
-}
-
-function seatsFilled(classPageInfo: GetClassPageInfoQuery): number[] {
-  return getValidSections(classPageInfo).map(
-    (section) => section.seatsCapacity - section.seatsRemaining
-  );
-}
-
-function seatsAvailable(classPageInfo: GetClassPageInfoQuery): number[] {
-  return getValidSections(classPageInfo).map(
-    (section) => section.seatsCapacity
-  );
-}
-
-// if there's at least one section with seatCapacity < 9999,
-// returns sections excluding those with seatCapacity = 9999
-function getValidSections(classPageInfo: GetClassPageInfoQuery) {
-  let allSections = classPageInfo.class.allOccurrences
-    .map((occurrence) => occurrence.sections)
-    .flat();
-  if (allSections.find((section) => section.seatsCapacity < 9999)) {
-    allSections = allSections.filter((section) => section.seatsCapacity < 9999);
-  }
-  return allSections;
-}
-
-function numberOfSections(classPageInfo: GetClassPageInfoQuery): number[] {
-  return classPageInfo.class.allOccurrences.map(
-    (occurrence) => occurrence.sections.length
-  );
-}
-
-type PrereqsDisplayProps = {
-  termId: string;
-  campus: string;
-  prereqs: Requisite;
-  indents: number;
-};
 
 function PrereqsDisplay({
   termId,
